fix mapper

+++ frontend/shared/sync/data/src/commonMain/kotlin/sync/repositories/SyncManagerImpl.kt
@@ -78,39 +78,41 @@ class SyncManagerImpl(
         Logg.debug { "Syncing start" }
         // TODO
         val unsyncedRows = queueRepo.getUnsynced()
-        val dtoOperations = unsyncedRows.map { mapSyncQueueRow(it) }
-        val request = SyncRequest(operations = dtoOperations, lastSyncDate = TODO())
-        val response = remoteDataSource.sendSyncRequest(request)
+        if (!unsyncedRows.isEmpty()) {
+            val dtoOperations = unsyncedRows.map { mapSyncQueueRow(it) }
+            val request = SyncRequest(operations = dtoOperations, lastSyncDate = TODO())
+            val response = remoteDataSource.sendSyncRequest(request)
 
-        queueRepo.withSyncDisabled {
-            if (response.acceptedIds.isNotEmpty()) {
-                queueRepo.deleteByProcessingIds(response.acceptedIds)
-            }
-            response.deleteOperations.forEach {
-               op ->
-                when(op.tableType) {
-                    SyncTableType.ACCOUNTS -> accountsRepo.softDeleteAccount(op.id)
-                    SyncTableType.CATEGORIES -> categoriesRepo.softDeleteCategory(op.id)
-                    else -> TODO()
+            queueRepo.withSyncDisabled {
+                if (response.acceptedIds.isNotEmpty()) {
+                    queueRepo.deleteByProcessingIds(response.acceptedIds)
                 }
-            }
-
-            response.updateState.forEach {
-                upd ->
-                val record = upd.getTypedRecord(json)
-                when(upd.tableType) {
-                    SyncTableType.ACCOUNTS -> {
-                        val accDto = record as AccountOutDTO
-                        val color = HexColor(hex = accDto?.color.orEmpty())
-                        accountsRepo.upsertAccount(id = accDto.id, name = accDto.name, color , stringAmount = accDto.funds)
+                response.deleteOperations.forEach {
+                        op ->
+                    when(op.tableType) {
+                        SyncTableType.ACCOUNTS -> accountsRepo.softDeleteAccount(op.id)
+                        SyncTableType.CATEGORIES -> categoriesRepo.softDeleteCategory(op.id)
+                        else -> TODO()
                     }
-                    SyncTableType.CATEGORIES -> {
-                        val catDto = record as CategoryRecordCreateDTO
-                        categoriesRepo.upsertCategory(id = catDto.id, name = catDto.name, emoji = catDto.emoji, type = catDto.type)
-                    }
-                    else -> TODO("ADD TRANSFERS AND OPERATIONS")
                 }
 
+                response.updateState.forEach {
+                        upd ->
+                    val record = upd.getTypedRecord(json)
+                    when(upd.tableType) {
+                        SyncTableType.ACCOUNTS -> {
+                            val accDto = record as AccountOutDTO
+                            val color = HexColor(hex = accDto?.color.orEmpty())
+                            accountsRepo.upsertAccount(id = accDto.id, name = accDto.name, color , stringAmount = accDto.funds)
+                        }
+                        SyncTableType.CATEGORIES -> {
+                            val catDto = record as CategoryRecordCreateDTO
+                            categoriesRepo.upsertCategory(id = catDto.id, name = catDto.name, emoji = catDto.emoji, type = catDto.type)
+                        }
+                        else -> TODO("ADD TRANSFERS AND OPERATIONS")
+                    }
+
+                }
             }
         }
